% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{gdc_palette}
\alias{gdc_palette}
\title{A Giorgio De Chirico color palette generator}
\usage{
gdc_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{PiazzaDItalia}, \code{LeMuseInquietanti}, \code{SoleSulCavalletto1972},
\code{PiazzaDItalia1939}, \code{ArrivoDelTrasloco1965}}

\item{n}{Number of colors desired. Unfortunately most palettes now only
have 4 or 5 colors. But hopefully we'll add more palettes soon.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes from Giorgio De Chirico paintings.
}
\examples{
gdc_palette("PiazzaDItalia")
gdc_palette("LeMuseInquietanti")
gdc_palette("SoleSulCavalletto1972")
gdc_palette("SoleSulCavalletto1972", 3)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- gdc_palette(21, name = "PiazzaDItalia", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
