% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{gdc_palettes}
\alias{gdc_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
gdc_palettes
}
\description{
Use \code{\link{gdc_palettes}} to construct palettes of desired length.
}
\keyword{datasets}
