% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallete_from_hex.R
\name{palette_from_hex}
\alias{palette_from_hex}
\title{Create a color palette from hex codes}
\usage{
palette_from_hex(hex_codes, name = "my_palette")
}
\arguments{
\item{hex_codes}{A character vector of hex color codes.}

\item{name}{A character string naming the palette.}
}
\value{
A list with class "artpalette" containing the name and colors
}
\description{
Create a color palette from hex codes
}
\examples{
palette_from_hex(c("#264653", "#2A9D8F", "#E9C46A"), name = "ocean")
}
