% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palette.R
\name{plot_palette}
\alias{plot_palette}
\title{Visualize a color palette as swatches}
\usage{
plot_palette(palette)
}
\arguments{
\item{palette}{An object of class "artpalette" from the function palette_from_hex()}
}
\value{
A ggplot2 object showing color swatches with hex labels
}
\description{
Visualize a color palette as swatches
}
\examples{
palette<-palette_from_hex(c("#264653", "#2A9D8F", "#E9C46A"), name = "ocean")
plot_palette(palette)
}
